Move files to correct location so we can actually load the gem..
