+++ jekyll-webmention_io.gemspec
@@ -1,7 +1,7 @@
 # coding: utf-8
 lib = File.expand_path('../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'jekyll/webmention_io/version'
+require 'jekyll-webmention_io/version'
 
 Gem::Specification.new do |spec|
   spec.name          = "jekyll-webmention_io"
